Ajustado create de eventos

+++ app/src/main/java/com/gaj2l/eventtus/models/Attachment.java
@@ -53,7 +53,7 @@ public class Attachment extends Entity {
         return activityId;
     }
 
-    public void setActivityId(int activityId) {
+    public void setActivityId(long activityId) {
         this.activityId = activityId;
     }
